+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Friday, May 24th, 2024
+
+### Added
+
+* Added a Messaging Info navigation menu to include links to key messaging resources
+* Added a Help/Feedback button linking to the #skylight-messaging-system slack channel
+
 ## Tuesday, May 21st, 2024
 
 ### Changed